+++ doc/development/release_notes.rst
@@ -71,6 +71,7 @@ Keep up-to-date with the latest changes to the Arcade library by the release not
 
   * :py:class:`~arcade.gui.UIBoxLayout` supports now align in constructor (changing later requires a `UIBoxLayout.trigger_full_render()`).
   * :py:class:`~arcade.gui.UIBoxLayout` supports now space_between in constructor.
+  * :py:class:`~arcade.gui.UIManager` fix #1067, consume press and release mouse events